refactor: use URL object instead of strings
